+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.11.27
+
+- Support parsing `GENERATED ALWAYS AS expr STORED` for psqldef [#184](https://github.com/k0kubun/sqldef/pull/184)
+- Support parsing `text_pattern_ops` for psqldef [#184](https://github.com/k0kubun/sqldef/pull/184)
+
+## v0.11.26
+
+- Support parsing REFERENCES .. ON DELETE/UPDATE on a column for psqldef [#184](https://github.com/k0kubun/sqldef/pull/184)
+
 ## v0.11.25
 
 - Fix schema handling of CREATE TABLE for psqldef [#187](https://github.com/k0kubun/sqldef/pull/187)